update top creators section
