+++ canister_ids.json
@@ -9,6 +9,6 @@
   },
   "vibeverse_token": {
     "ic": "gqljf-taaaa-aaaap-abdea-cai",
-    "staging": "q5xlo-wyaaa-aaaak-afciq-cai"
+    "staging": "xufih-jyaaa-aaaak-afjia-cai"
   }
 }
