+++ vibeverse_backend/backend/Cargo.toml
@@ -17,7 +17,7 @@ serde = { workspace = true }
 serde_bytes = { workspace = true }
 tracing = { workspace = true }
 async-trait = "0.1.77"
-
+itertools = "0.12.0"
 libraries = { path = "../../libraries" }
 
 [dev-dependencies]
