+++ vibeverse_backend/backend/src/creators.rs
@@ -1,7 +1,9 @@
 use crate::{
     memory::CREATORS,
     types::{Creator, StorableNat, UserId},
+    Badge,
 };
+use itertools::Itertools;
 
 /// Sets the metadata for the specific creator.
 pub fn set_creator_metadata(user_id: UserId, creator: Creator) -> Result<(), String> {
@@ -23,4 +25,18 @@ pub fn add_score(user_id: UserId, score: u8) -> Result<StorableNat, String> {
     } else {
         Ok(Default::default())
     }
+}
+
+pub fn top_n_creators(n: usize) -> Vec<(UserId, Creator, Badge)> {
+    CREATORS.with(|creators| {
+        creators
+            .borrow()
+            .iter()
+            .map(|(user_id, creator)| (Into::<UserId>::into(user_id.clone()), creator.clone(), creator.badge()))
+            // sort by score
+            .sorted_by(|(_, a, _), (_, b, _)| b.score.cmp(&a.score))
+            // take top n
+            .take(n)
+            .collect()
+    })
 }
